Switch from using includes to defining lots of blocks. See additional lines of this message for more info.
From: https://docs.djangoproject.com/en/2.2/ref/templates/builtins/#include
> Blocks are evaluated before they are included. This means that a template that includes blocks from another will contain blocks that have already been evaluated and rendered - not blocks that can be overridden by, for example, an extending template.

This was preventing overriding the page title from working, and was making it hard to add additional JS to a particular page.

Meanwhile, blocks can apparently call super to extend existing blocks.

https://stackoverflow.com/questions/8717521/django-multi-level-template-extends-and-nested-blocks

So we'll just have everything in the base templates with places we want to be customizable specified as blocks.